+++ spec/nokodiff_spec.rb
@@ -5,60 +5,6 @@ RSpec.describe Nokodiff do
     expect(Nokodiff::VERSION).not_to be nil
   end
 
-  describe ".diff" do
-    it "allows nil as an input" do
-      expect {
-        Nokodiff.diff("<p>html snippet</p>", nil)
-      }.not_to raise_error
-    end
-
-    it "allows '' as an input" do
-      expect {
-        Nokodiff.diff("<p>html snippet</p>", "")
-      }.not_to raise_error
-    end
-
-    it "allows HTML comments within an input" do
-      expect {
-        Nokodiff.diff("<!-- hello --><p>html snippet</p>", "")
-      }.not_to raise_error
-    end
-
-    it "raises an argument error if only one block is given" do
-      expect {
-        Nokodiff.diff("just text")
-      }.to raise_error(ArgumentError)
-    end
-
-    it "does not raise an argument error when two blocks are given" do
-      expect {
-        Nokodiff.diff("<p>block one</p>", "<p>block two</p>")
-      }.not_to raise_error
-    end
-
-    it "raises an argument error when passed non html arguments" do
-      expect {
-        Nokodiff.diff("just text", "<p>html snippet</p>")
-      }.to raise_error(ArgumentError)
-    end
-
-    it "raises an argument error when passed malformed HTML" do
-      invalid_html = "<<p> /p>"
-
-      expect {
-        Nokodiff.diff(invalid_html, "<p>html snippet</p>")
-      }.to raise_error(ArgumentError)
-    end
-
-    it "raises an argument error when passed preprocessing instructions" do
-      invalid_html = '<?xml version="1.0"?><div></div>'
-
-      expect {
-        Nokodiff.diff(invalid_html, "<p>html snippet</p>")
-      }.to raise_error(ArgumentError)
-    end
-  end
-
   describe ".safe_html" do
     before { stub_const("Differ", Class.new) }
 
@@ -92,10 +38,10 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(html, html)
 
-          expect(result).to include("<p>Hello world!</p>")
-          expect(result).not_to include('<div class="diff">')
-          expect(result).not_to include('<del aria-label="removed content">')
-          expect(result).not_to include('<ins aria-label="added content">')
+          expect(result).to have_tag("p", text: "Hello world!")
+          expect(result).not_to have_tag("div", with: { class: "diff" })
+          expect(result).not_to have_tag("del", with: { "aria-label" => "removed content" })
+          expect(result).not_to have_tag("ins", with: { "aria-label" => "added content" })
         end
       end
 
@@ -106,9 +52,18 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(before_html, after_html)
 
-          expect(result).to include('<div class="diff">')
-          expect(result).to include('<del aria-label="removed content"><p><strong>Hell</strong>o world!</p></del>')
-          expect(result).to include('<ins aria-label="added content"><p><strong>G</strong>o<strong>odbye</strong> world!</p></ins>')
+          expect(result).to have_tag("div", with: { class: "diff" })
+          expect(result).to have_tag("del", with: { "aria-label" => "removed content" }) do
+            with_tag("p") do
+              with_tag("strong", text: "Hell")
+            end
+          end
+          expect(result).to have_tag("ins", with: { "aria-label" => "added content" }) do
+            with_tag("p") do
+              with_tag("strong", text: "G")
+              with_tag("strong", text: "odbye")
+            end
+          end
         end
       end
 
@@ -119,9 +74,11 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(before_html, after_html)
 
-          expect(result).to include('<div class="diff">')
-          expect(result).to include('<del aria-label="removed content"><p>Hello world!</p></del>')
-          expect(result).not_to include('<ins aria-label="added content">')
+          expect(result).to have_tag("div", with: { class: "diff" })
+          expect(result).to have_tag("del", with: { "aria-label" => "removed content" }) do
+            with_tag("p", text: "Hello world!")
+          end
+          expect(result).not_to have_tag("ins", with: { "aria-label" => "added content" })
         end
       end
 
@@ -132,12 +89,15 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(before_html, after_html)
 
-          expect(result).to include('<div class="diff">')
-          expect(result).not_to include('<del aria-label="removed content">')
-          expect(result).to include('<ins aria-label="added content"><p>Hello world!</p></ins>')
+          expect(result).to have_tag("div", with: { class: "diff" })
+          expect(result).not_to have_tag("del", with: { "aria-label" => "removed content" })
+          expect(result).to have_tag("ins", with: { "aria-label" => "added content" }) do
+            with_tag("p", text: "Hello world!")
+          end
         end
       end
     end
+
     context "links" do
       it "diffs changed link text" do
         before_html = <<~HTML
@@ -160,11 +120,14 @@ RSpec.describe Nokodiff do
 
         output = Nokodiff.diff(before_html, after_html)
 
-        expect(output).to include('<li><a href="https://a.example.com">Link <strong>A</strong></a></li>')
-        expect(output).to include('<li><a href="https://a.example.com">Link <strong>B</strong></a></li>')
-
-        expect(output).to include('<del aria-label="removed content">')
-        expect(output).to include('<ins aria-label="added content">')
+        expect(output).to have_tag("a", with: { href: "https://a.example.com" }) do
+          with_tag("strong", text: "A")
+        end
+        expect(output).to have_tag("a", with: { href: "https://a.example.com" }) do
+          with_tag("strong", text: "B")
+        end
+        expect(output).to have_tag("del", with: { "aria-label" => "removed content" })
+        expect(output).to have_tag("ins", with: { "aria-label" => "added content" })
       end
 
       it "diffs a removed link against the matching line" do
@@ -189,9 +152,15 @@ RSpec.describe Nokodiff do
 
         output = Nokodiff.diff(before_html, after_html)
 
-        expect(output).to include('<li><a href="https://a.example.com">Link <strong>A</strong></a></li>')
-        expect(output).to include('<li><a href="https://b.example.com"><strong>Link B</strong></a></li>')
-        expect(output).to include('<li><a href="https://b.example.com">Link <strong>B</strong></a></li>')
+        expect(output).to have_tag("a", with: { href: "https://a.example.com" }) do
+          with_tag("strong", text: "A")
+        end
+        expect(output).to have_tag("a", with: { href: "https://b.example.com" }) do
+          with_tag("strong", text: "Link B")
+        end
+        expect(output).to have_tag("a", with: { href: "https://b.example.com" }) do
+          with_tag("strong", text: "B")
+        end
       end
     end
 
@@ -203,8 +172,12 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(before_html, after_html)
 
-          expect(result).to include('<div class="diff">')
-          expect(result).to include('<ins aria-label="added content"><p> a <strong>b c</strong></p></ins>')
+          expect(result).to have_tag("div", with: { class: "diff" })
+          expect(result).to have_tag("ins", with: { "aria-label" => "added content" }) do
+            with_tag("p") do
+              with_tag("strong", text: "b c")
+            end
+          end
         end
       end
 
@@ -215,8 +188,58 @@ RSpec.describe Nokodiff do
 
           result = Nokodiff.diff(before_html, after_html)
 
-          expect(result).to include('<div class="diff">')
-          expect(result).to include('<ins aria-label="added content"><p> <strong>a </strong>b <strong>c</strong></p></ins>')
+          expect(result).to have_tag("div", with: { class: "diff" })
+          expect(result).to have_tag("ins", with: { "aria-label" => "added content" }) do
+            with_tag("p") do
+              with_tag("strong", text: "a ")
+              with_tag("strong", text: "c")
+            end
+          end
+        end
+      end
+    end
+  end
+
+  context "when `data-diff-key` attributes are present" do
+    context "when an element has been added" do
+      let(:before_html) { load_fixture("complex/added/before") }
+      let(:after_html) { load_fixture("complex/added/after") }
+
+      it "adds a diff showing the added content" do
+        result = Nokodiff.diff(before_html, after_html)
+
+        expect(result).to have_tag("div", with: { "data-diff-key" => "description" }) do
+          with_tag("div", class: "diff") do
+            with_tag("ins", with: { "aria-label" => "added content" }) do
+              with_text("Main contact info")
+            end
+            without_tag("del")
+          end
+        end
+
+        expect(result).to have_tag("div", with: { "data-diff-key" => "telephone-1" }) do
+          without_tag("ins")
+          without_tag("del")
+        end
+      end
+    end
+
+    context "when an element has been modified" do
+      let(:before_html) { load_fixture("complex/modified/before") }
+      let(:after_html) { load_fixture("complex/modified/after") }
+
+      it "adds a diff showing the content modifications" do
+        result = Nokodiff.diff(before_html, after_html)
+
+        expect(result).to have_tag("div", with: { "data-diff-key" => "telephone-1" }) do
+          with_tag("div", class: "diff") do
+            with_tag("del", with: { "aria-label" => "removed content" }) do
+              with_tag("span", class: "tel", text: "0300 123 123")
+            end
+            with_tag("ins", with: { "aria-label" => "added content" }) do
+              with_tag("span", class: "tel", text: "0300 345 345")
+            end
+          end
         end
       end
     end